+++ src/lib.rs
@@ -6,6 +6,7 @@ pub mod commands;
 pub mod config;
 pub mod error;
 pub mod project;
+pub mod resources;
 pub mod scripts;
 pub mod update_check;
 pub mod utils;